Moving some cases from ClipboardTests class to ClipboardCoreTests
